+++ androidplot-core/src/main/java/com/androidplot/xy/StepFormatter.java
@@ -36,7 +36,7 @@ public class StepFormatter extends LineAndPointFormatter {
     }
 
     @Override
-    public SeriesRenderer getRendererInstance(XYPlot plot) {
+    public SeriesRenderer doGetRendererInstance(XYPlot plot) {
         return new StepRenderer(plot);
     }