+++ androidplot-core/src/test/java/com/androidplot/xy/BubbleRendererTest.java
@@ -43,14 +43,15 @@ public class BubbleRendererTest extends AndroidplotTest {
 
     @Before
     public void setUp() throws Exception {
-        canvas = new Canvas();
+        canvas = spy(new Canvas());
         xyPlot = spy(new XYPlot(getContext(), "My Plot"));
     }
 
     @Test
     public void testOnRender() throws Exception {
         BubbleFormatter formatter = spy(new BubbleFormatter());
-        BubbleRenderer renderer = spy((BubbleRenderer) formatter.getRendererInstance(xyPlot));
+        BubbleRenderer br = formatter.getRendererInstance(xyPlot);
+        BubbleRenderer renderer = spy(br);
 
         doReturn(renderer.getClass()).when(formatter).getRendererClass();
         doReturn(renderer).when(formatter).getRendererInstance(any(XYPlot.class));
@@ -60,5 +61,26 @@ public class BubbleRendererTest extends AndroidplotTest {
         xyPlot.addSeries(bs, formatter);
 
         renderer.onRender(canvas, plotArea, bs, formatter, renderStack);
+    }
+
+    @Test
+    public void testOnRender_withPointLabeler() throws Exception {
+        BubbleFormatter formatter = spy(new BubbleFormatter());
+        PointLabelFormatter plf = new PointLabelFormatter(Color.MAGENTA);
+        formatter.setPointLabelFormatter(plf);
+        BubbleRenderer br = formatter.getRendererInstance(xyPlot);
+        BubbleRenderer renderer = spy(br);
+
+        doReturn(renderer.getClass()).when(formatter).getRendererClass();
+        doReturn(renderer).when(formatter).getRendererInstance(any(XYPlot.class));
+
+        BubbleSeries bs = new BubbleSeries(2, 2, 22);
+
+        xyPlot.addSeries(bs, formatter);
+
+        renderer.onRender(canvas, plotArea, bs, formatter, renderStack);
+
+        // verify the z-val is the one labeled:
+        verify(canvas).drawText(eq("22"), anyFloat(), anyFloat(), eq(plf.getTextPaint()));
     }
 }